Some todos (#26)
* better fallback libero handling.

It does not show the `CL` option (Change Libero) when a fallback libero
is not available.

* better import/export fs dialog.

You now select directories with 'Space' and confirm selection by 'Enter'.
Adding a header showing current location.

* introducing fs abstractions.

By introducing proper abstractions for fs we can improve testability and
do not couple the application itself to work with file system.
A couple of minor bug fixes.

* localization improvements.

* switch to AGPL-3.0-only licensing.

* review match reporting.